+++ .gitignore
@@ -136,4 +136,5 @@ database.db
 neuroAPI/logs
 famousblocks1mLesnoe.csv
 test.pt
+trash/
 # -- omelched
