+++ neuroAPI/neuralmodule/__init__.py
@@ -1,12 +1,1 @@
-import torch
-from torch import nn
 
-from neuroAPI.neuralmodule.network import NeuralNetwork
-model = NeuralNetwork(3)
-
-learning_rate = 1e-3
-batch_size = 64
-epochs = 1000
-
-loss_fn = nn.MSELoss()
-optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
